test policies on simple bank account

+++ tests/test_simple_bank_account.py
@@ -1,8 +1,12 @@
 import unittest
 
+from unittest.mock import patch, Mock
+
 import os
 
 from dfm18.bank_account._simple import SimpleBankAccount
+
+from dfm18.bank_account.policies import Policy
 
 
 class TestSimpleBankAccount(unittest.TestCase):
@@ -13,13 +17,37 @@ class TestSimpleBankAccount(unittest.TestCase):
         for handler in self.account._logger.handlers:
             handler.close()
 
-        if os.path.exists(self.account.log_file):
+        if self.account.log_file is not None and os.path.exists(self.account.log_file):
             os.remove(self.account.log_file)
+
+    @patch("dfm18.bank_account.policies.Policy")
+    def test_initialization_with_policies(self, mock_policy: Mock):
+        mock_policy_1 = Mock(spec=Policy)
+        mock_policy_2 = Mock(spec=Policy)
+
+        mock_policy_1.apply.return_value = None
+        mock_policy_1.supports.return_value = True
+        mock_policy_2.apply.return_value = None
+        mock_policy_2.supports.return_value = False
+
+        policies = [mock_policy_1, mock_policy_2]
+
+        self.account = SimpleBankAccount(balance=1000, policies=policies)
+
+        self.assertEqual(len(self.account.policies), 2)
+        self.assertIn(mock_policy_1, self.account.policies)
+        self.assertIn(mock_policy_2, self.account.policies)
+        self.assertEqual(self.account.policies, policies)
+
+        mock_policy_1.apply.assert_not_called()
+        mock_policy_1.supports.assert_not_called()
+        mock_policy_2.apply.assert_not_called()
+        mock_policy_2.supports.assert_not_called()
 
     def test_deposit(self):
         new_balance = self.account.deposit(500)
         self.assertEqual(new_balance, 1500)
-    
+
     def test_deposit_on_negative_amount_raises_exception(self):
         with self.assertRaises(ValueError):
             self.account.deposit(-100)
@@ -28,10 +56,58 @@ class TestSimpleBankAccount(unittest.TestCase):
         with self.assertRaises(ValueError):
             self.account.deposit(0)
 
+    @patch("dfm18.bank_account.policies.Policy")
+    def test_deposit_on_policy_violation_raises_exception(self, mock_policy: Mock):
+        mock_policy.apply.side_effect = Exception("Policy Restriction")
+        mock_policy.supports.return_value = True
+
+        self.account = SimpleBankAccount(balance=1000, policies=[mock_policy])
+
+        with self.assertRaises(Exception) as context:
+            self.account.deposit(100)
+
+        self.assertEqual(str(context.exception), "Policy Restriction")
+
+    @patch("dfm18.bank_account.policies.Policy")
+    def test_deposit_allowed_by_policy(self, mock_policy: Mock):
+        mock_policy.apply.return_value = None
+        mock_policy.supports.return_value = True
+
+        self.account = SimpleBankAccount(balance=1000, policies=[mock_policy])
+
+        amount = 100
+
+        try:
+            self.account.deposit(amount)
+        except Exception as e:
+            self.fail(f"Unexepected exception: {str(e)}")
+
+        mock_policy.apply.assert_called_once_with(self.account, amount)
+        mock_policy.supports.assert_called_once()
+
+    @patch("dfm18.bank_account.policies.Policy")
+    def test_deposit_allowed_due_to_missing_policies_for_operation(
+        self, mock_policy: Mock
+    ):
+        mock_policy.apply.return_value = None
+        mock_policy.supports.return_value = False
+
+        self.account = SimpleBankAccount(balance=1000, policies=[mock_policy])
+
+        amount = 100
+
+        try:
+            self.account.deposit(amount)
+        except Exception as e:
+            self.fail(f"Unexepected exception: {str(e)}")
+
+        mock_policy.apply.assert_not_called()
+        mock_policy.supports.assert_called_once()
+
     def test_withdraw(self):
         new_balance = self.account.withdraw(200)
         self.assertEqual(new_balance, 800)
-    
+
     def test_withdraw_on_negative_amount_raises_exception(self):
         with self.assertRaises(ValueError):
             self.account.withdraw(-100)
@@ -39,6 +115,54 @@ class TestSimpleBankAccount(unittest.TestCase):
     def test_withdraw_on_zero_amount_raises_exception(self):
         with self.assertRaises(ValueError):
             self.account.withdraw(0)
+
+    @patch("dfm18.bank_account.policies.Policy")
+    def test_withdraw_on_policy_violation_raises_exception(self, mock_policy: Mock):
+        mock_policy.apply.side_effect = Exception("Policy Restriction")
+        mock_policy.supports.return_value = True
+
+        self.account = SimpleBankAccount(balance=1000, policies=[mock_policy])
+
+        with self.assertRaises(Exception) as context:
+            self.account.withdraw(100)
+
+        self.assertEqual(str(context.exception), "Policy Restriction")
+
+    @patch("dfm18.bank_account.policies.Policy")
+    def test_withdraw_allowed_by_policy(self, mock_policy: Mock):
+        mock_policy.apply.return_value = None
+        mock_policy.supports.return_value = True
+
+        self.account = SimpleBankAccount(balance=1000, policies=[mock_policy])
+
+        amount = 100
+
+        try:
+            self.account.withdraw(amount)
+        except Exception as e:
+            self.fail(f"Unexepected exception: {str(e)}")
+
+        mock_policy.apply.assert_called_once_with(self.account, amount)
+        mock_policy.supports.assert_called_once()
+
+    @patch("dfm18.bank_account.policies.Policy")
+    def test_withdraw_allowed_due_to_missing_policies_for_operation(
+        self, mock_policy: Mock
+    ):
+        mock_policy.apply.return_value = None
+        mock_policy.supports.return_value = False
+
+        self.account = SimpleBankAccount(balance=1000, policies=[mock_policy])
+
+        amount = 100
+
+        try:
+            self.account.withdraw(amount)
+        except Exception as e:
+            self.fail(f"Unexepected exception: {str(e)}")
+
+        mock_policy.apply.assert_not_called()
+        mock_policy.supports.assert_called_once()
 
     def test_balance_property(self):
         self.assertEqual(self.account.balance, 1000)
